+++ lib/run_app_with_flavor.dart
@@ -21,20 +21,17 @@ import 'package:package_info/package_info.dart';
 import 'package:provider/provider.dart';
 import 'package:sentry/sentry.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:wr_app/domain/lesson/index.dart';
-import 'package:wr_app/domain/system/index.dart';
-import 'package:wr_app/domain/user/index.dart';
 import 'package:wr_app/infrastructure/api/functions.dart';
-import 'package:wr_app/infrastructure/article/article_persistence.dart';
-import 'package:wr_app/infrastructure/article/article_persistence_mock.dart';
-import 'package:wr_app/infrastructure/auth/auth_persistence.dart';
-import 'package:wr_app/infrastructure/lesson/lesson_persistence.dart';
-import 'package:wr_app/infrastructure/lesson/lesson_persistence_mock.dart';
-import 'package:wr_app/infrastructure/note/note_persistence.dart';
-import 'package:wr_app/infrastructure/shop/shop_persistence.dart';
-import 'package:wr_app/infrastructure/shop/shop_persistence_mock.dart';
-import 'package:wr_app/infrastructure/system/system_persistence.dart';
-import 'package:wr_app/infrastructure/user/user_persistence.dart';
+import 'package:wr_app/infrastructure/article/article_repository.dart';
+import 'package:wr_app/infrastructure/article/article_repository_mock.dart';
+import 'package:wr_app/infrastructure/auth/auth_repository.dart';
+import 'package:wr_app/infrastructure/lesson/lesson_repository.dart';
+import 'package:wr_app/infrastructure/lesson/lesson_repository_mock.dart';
+import 'package:wr_app/infrastructure/note/note_repository.dart';
+import 'package:wr_app/infrastructure/shop/shop_repository.dart';
+import 'package:wr_app/infrastructure/shop/shop_repository_mock.dart';
+import 'package:wr_app/infrastructure/system/system_repository.dart';
+import 'package:wr_app/infrastructure/user/user_repository.dart';
 import 'package:wr_app/presentation/app.dart';
 import 'package:wr_app/presentation/article/notifier/article_notifier.dart';
 import 'package:wr_app/presentation/auth_notifier.dart';
