+++ test/domain/user_test.dart
@@ -12,7 +12,7 @@ import 'package:wr_app/domain/lesson/model/test_result.dart';
 import 'package:wr_app/domain/note/model/note.dart';
 import 'package:wr_app/domain/user/index.dart';
 import 'package:wr_app/infrastructure/api/functions.dart';
-import 'package:wr_app/infrastructure/user/user_persistence.dart';
+import 'package:wr_app/infrastructure/user/user_repository.dart';
 import 'package:wr_app/usecase/user_service.dart';
 
 extension StoreEx on FirebaseFirestore {
