+++ test/infrastructure/note_test.dart
@@ -1,7 +1,5 @@
 import 'package:cloud_firestore_mocks/cloud_firestore_mocks.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:wr_app/domain/user/index.dart';
-import 'package:wr_app/infrastructure/note/note_persistence.dart';
+import 'package:wr_app/infrastructure/note/note_repository.dart';
 
 void main() {
   final store = MockFirestoreInstance();
